Refactoring, work in progress
